requirements.txt
pathway
requests
pandas
tqdm
sentence-transformers
torch
python-dotenv
